Time: 2 ms (68.66%), Space: 44.9 MB (30.58%) - LeetHub

+++ 0152-maximum-product-subarray/0152-maximum-product-subarray.java
@@ -2,14 +2,14 @@ class Solution {
     public int maxProduct(int[] nums) {
         int n = nums.length; 
 
-        double pre = 1, suff = 1;
+        int pre = 1, suff = 1;
         int max = Integer.MIN_VALUE;
         for (int i = 0; i < n; i++) {
             if (pre == 0) pre = 1;
             if (suff == 0) suff = 1;
             pre *= (double)nums[i];
             suff *= (double)nums[n - i - 1];
-            max = (int)Math.max(max, Math.max(pre, suff));
+            max = Math.max(max, Math.max(pre, suff));
         }
         return max;
     }
